backend/main.py
from fastapi import FastAPI, UploadFile, File, HTTPException, Form
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import os
import json
import chardet
import asyncio
import google.generativeai as genai
from typing import Optional
import io

# Try to import document parsing libraries
try:
    import pdfplumber
    PDF_SUPPORT = True
except ImportError:
    PDF_SUPPORT = False

try:
    from docx import Document
    DOCX_SUPPORT = True
except ImportError:
    DOCX_SUPPORT = False

app = FastAPI(title="COBOL Code Analyzer API")

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # In production, replace with specific origins
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Initialize Gemini API
api_key = os.environ.get("GEMINI_API_KEY")
if api_key:
    genai.configure(api_key=api_key)
else:
    print("WARNING: GEMINI_API_KEY environment variable not set")

# Models
class CodeAnalysisRequest(BaseModel):
    code: str

class CodeAnalysisResponse(BaseModel):
    metrics: dict
    complexity: dict
    explanation: str
    chartData: dict

# Helper functions
def extract_text_from_pdf(file_content):
    if not PDF_SUPPORT:
        raise HTTPException(status_code=400, detail="PDF support not available. Install pdfplumber.")
    
    text = ""
    with pdfplumber.open(io.BytesIO(file_content)) as pdf:
        for page in pdf.pages:
            text += page.extract_text() or ""
    return text

def extract_text_from_docx(file_content):
    if not DOCX_SUPPORT:
        raise HTTPException(status_code=400, detail="DOCX support not available. Install python-docx.")
    
    doc = Document(io.BytesIO(file_content))
    text = "\n".join([paragraph.text for paragraph in doc.paragraphs])
    return text

def extract_text_from_file(file_content, file_extension):
    if file_extension in [".pdf"]:
        return extract_text_from_pdf(file_content)
    elif file_extension in [".docx", ".doc"]:
        return extract_text_from_docx(file_content)
    else:
        # For text files, detect encoding
        result = chardet.detect(file_content)
        encoding = result["encoding"] if result["encoding"] else "utf-8"
        try:
            return file_content.decode(encoding)
        except UnicodeDecodeError:
            return file_content.decode("latin-1")

def extract_metrics(code):
    """Extract basic metrics from code content"""
    lines = code.split("\n")
    loc = len([line for line in lines if line.strip()])
    
    # Count conditional blocks
    if_else_blocks = len([line for line in lines if any(keyword in line.upper() for keyword in ["IF", "ELSE", "WHEN", "EVALUATE"])])
    
    # Count variables (simplified - looking for level numbers in COBOL)
    variable_count = len([line for line in lines if line.strip() and line.strip()[0].isdigit()])
    
    # Estimate nested depth
    nested_depth = 0
    current_depth = 0
    for line in lines:
        line_upper = line.upper()
        if any(keyword in line_upper for keyword in ["IF", "EVALUATE", "PERFORM"]) and "END-" not in line_upper:
            current_depth += 1
            nested_depth = max(nested_depth, current_depth)
        elif any(keyword in line_upper for keyword in ["END-IF", "END-EVALUATE", "END-PERFORM"]):
            current_depth = max(0, current_depth - 1)
    
    return {
        "loc": loc,
        "ifElseBlocks": if_else_blocks,
        "variableCount": variable_count,
        "nestedDepth": nested_depth
    }

async def analyze_with_gemini(code):
    """Analyze code using Gemini API"""
    if not api_key:
        # Fallback if no API key
        return fallback_analysis(code)
    
    try:
        model = genai.GenerativeModel('gemini-pro')
        prompt = f"""
        You are a COBOL code analysis expert. Analyze the following code and provide:
        1. A brief explanation of what this code does (max 3 sentences).
        2. Classify the complexity as Simple, Moderate, or Complex.
        3. Provide a confidence score for your classification (0-100).
        4. List key factors that influenced your classification.
        
        Code to analyze:
        ```
        {code[:8000]}  # Limit code length
        ```
        
        Format your response in plain text with clear sections.
        """
        
        response = await asyncio.to_thread(
            lambda: model.generate_content(prompt).text
        )
        
        # Parse the response
        explanation = response[:500]  # First 500 chars as explanation
        
        # Extract classification
        classification = "Moderate"  # Default
        if "simple" in response.lower():
            classification = "Simple"
        elif "complex" in response.lower():
            classification = "Complex"
        
        # Extract confidence score if present
        confidence_match = response.lower().find("confidence")
        confidence_score = 75.0  # Default
        if confidence_match != -1:
            # Try to find a number after "confidence"
            text_after = response[confidence_match:confidence_match+30]
            import re
            score_match = re.search(r'\d+', text_after)
            if score_match:
                confidence_score = float(score_match.group())
                confidence_score = min(100.0, max(0.0, confidence_score))  # Ensure in range 0-100
        
        return {
            "classification": classification,
            "confidenceScore": confidence_score,
            "explanation": explanation
        }
        
    except Exception as e:
        print(f"Error calling Gemini API: {str(e)}")
        return fallback_analysis(code)

def fallback_analysis(code):
    """Fallback analysis when Gemini API is unavailable"""
    metrics = extract_metrics(code)
    
    # Simple scoring system
    complexity_score = 0
    
    # Lines of code contribution
    if metrics["loc"] < 100: complexity_score += 1
    elif metrics["loc"] < 300: complexity_score += 2
    elif metrics["loc"] < 500: complexity_score += 3
    else: complexity_score += 4
    
    # Conditional blocks contribution
    if metrics["ifElseBlocks"] < 10: complexity_score += 1
    elif metrics["ifElseBlocks"] < 25: complexity_score += 2
    elif metrics["ifElseBlocks"] < 50: complexity_score += 3
    else: complexity_score += 4
    
    # Variable count contribution
    if metrics["variableCount"] < 20: complexity_score += 1
    elif metrics["variableCount"] < 50: complexity_score += 2
    elif metrics["variableCount"] < 100: complexity_score += 3
    else: complexity_score += 4
    
    # Nested depth contribution
    if metrics["nestedDepth"] < 3: complexity_score += 1
    elif metrics["nestedDepth"] < 5: complexity_score += 2
    elif metrics["nestedDepth"] < 7: complexity_score += 3
    else: complexity_score += 4
    
    # Classification based on total score
    if complexity_score <= 6:
        classification = "Simple"
        confidence_score = 85 + (hash(code) % 10)  # Deterministic random-like number
    elif complexity_score <= 10:
        classification = "Moderate"
        confidence_score = 75 + (hash(code) % 15)
    else:
        classification = "Complex"
        confidence_score = 80 + (hash(code) % 15)
    
    return {
        "classification": classification,
        "confidenceScore": confidence_score,
        "explanation": "Analysis performed using metrics-based classification. Enable Gemini API for more detailed analysis."
    }

@app.get("/")
async def root():
    return {"message": "COBOL Code Analyzer API is running"}

@app.post("/analyze", response_model=CodeAnalysisResponse)
async def analyze_code(request: CodeAnalysisRequest):
    """Analyze code content directly"""
    if not request.code or not isinstance(request.code, str):
        raise HTTPException(status_code=400, detail="Invalid request: Code content is required")
    
    try:
        # Extract metrics
        metrics = extract_metrics(request.code)
        
        # Get Gemini analysis
        gemini_analysis = await analyze_with_gemini(request.code)
        
        # Prepare chart data
        chart_data = {
            "labels": ["Lines of Code", "IF/ELSE Blocks", "Variables", "Nested Depth"],
            "datasets": [{
                "label": "Code Metrics",
                "data": [metrics["loc"], metrics["ifElseBlocks"], metrics["variableCount"], metrics["nestedDepth"]],
                "backgroundColor": [
                    "rgba(54, 162, 235, 0.6)",
                    "rgba(255, 206, 86, 0.6)",
                    "rgba(75, 192, 192, 0.6)",
                    "rgba(153, 102, 255, 0.6)"
                ],
                "borderColor": [
                    "rgba(54, 162, 235, 1)",
                    "rgba(255, 206, 86, 1)",
                    "rgba(75, 192, 192, 1)",
                    "rgba(153, 102, 255, 1)"
                ],
                "borderWidth": 1
            }]
        }
        
        return {
            "metrics": metrics,
            "complexity": {
                "classification": gemini_analysis["classification"],
                "confidenceScore": gemini_analysis["confidenceScore"]
            },
            "explanation": gemini_analysis["explanation"],
            "chartData": chart_data
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error analyzing code: {str(e)}")

@app.post("/upload")
async def upload_file(file: UploadFile = File(...)):
    """Upload and analyze a file"""
    try:
        # Get file extension
        filename = file.filename
        file_extension = os.path.splitext(filename)[1].lower()
        
        # Read file content
        file_content = await file.read()
        
        # Extract text based on file type
        try:
            text_content = extract_text_from_file(file_content, file_extension)
        except Exception as e:
            raise HTTPException(status_code=400, detail=f"Error extracting text from file: {str(e)}")
        
        # Analyze the extracted text
        request = CodeAnalysisRequest(code=text_content)
        return await analyze_code(request)
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error processing file: {str(e)}")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)
