backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chardet==5.2.0
google-generativeai==0.3.1
pdfplumber==0.10.2
python-docx==1.0.1
asyncio==3.4.3
pydantic==2.4.2